--- plugin/web/extensions/sidebar/popups/BasicOperation/index.html
@@ -1,19 +1,0 @@
-<title></title>
-<style>
- html, body{ background: transparent; font-family: 'Roboto', sans-serif; } 
- .root{ background: #fff;
-border-radius: 0 0 8px 0; }
-
-  html,
-  body,
-  .root {
-    width: 350px;
-    height: 680px;
-  }
-  .extended{ height: 100%; }
-
-</style>
-<body>
-  <div class="root" id="root"></div>
-</body>
-<script type="module" src="./main.tsx"></script>
