--- plugin/web/extensions/sidebar/popups/ShadowFunction/main.tsx
@@ -1,13 +1,0 @@
-import ReactDOM from "react-dom/client";
-
-import ShadowFunction from ".";
-
-(async () => {
-  const element = document.getElementById("root");
-  if (element) {
-    const root = ReactDOM.createRoot(element);
-    root.render(<ShadowFunction />);
-  }
-})();
-
-export {};
