--- plugin/web/extensions/sidebar/popups/TryMapInfo/index.tsx
@@ -1,96 +1,0 @@
-import mapVideo from "@web/extensions/sidebar/core/assets/mapVideo.png";
-import { styled } from "@web/theme";
-
-import useGlobalHooks from "../globalHooks";
-import { NumberingWrapper, ParagraphItem, PopupWrapper } from "../sharedComponent";
-
-const TryMapInfo: React.FC = () => {
-  const { handleClosePopup, handleShowMapModal } = useGlobalHooks();
-
-  return (
-    <PopupWrapper handleClose={handleClosePopup}>
-      <Wrapper>
-        <Title>マップを使ってみる</Title>
-
-        <ContentWrapper>
-          <ImgWrapper>
-            <img src={mapVideo} onClick={handleShowMapModal} />
-          </ImgWrapper>
-          <Paragraph>
-            上の動画では、空間データや地図を扱うのが初めての方が、データを追加したり地図上に表現するために必要な、基本的な機能について紹介しています。
-            動画を見る時間がない方は、次の手順をお試しください。
-          </Paragraph>
-          <ParagraphItem>
-            <NumberingWrapper number={1} />
-            <Paragraph>
-              Add Data から使用可能なデータを表示して、マップに追加してみましょう。
-            </Paragraph>
-          </ParagraphItem>
-          <ParagraphItem>
-            <NumberingWrapper number={2} />
-            <Paragraph>
-              画面の左側に現れるボタンで表示/非表示ボタンを切り替えてみましょう。
-            </Paragraph>
-          </ParagraphItem>
-          <ParagraphItem>
-            <NumberingWrapper number={3} />
-            <Paragraph>
-              マップ上でデータをクリックして、より詳細な情報や元データについての情報を見てみましょう。
-            </Paragraph>
-          </ParagraphItem>
-          <ParagraphItem>
-            <NumberingWrapper number={4} />
-            <Paragraph>Map Settingsボタンをクリックして、背景図を変更してみましょう。</Paragraph>
-          </ParagraphItem>
-          <ParagraphItem>
-            <NumberingWrapper number={5} />
-            <Paragraph>
-              画面右手のズームや回転ボタンを使って、視点を変化させてみましょう。
-            </Paragraph>
-          </ParagraphItem>
-        </ContentWrapper>
-      </Wrapper>
-    </PopupWrapper>
-  );
-};
-
-export default TryMapInfo;
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 0px 16px;
-  gap: 24px;
-  width: 333px;
-  height: 712px;
-`;
-
-const ImgWrapper = styled.div`
-  width: 301px;
-  height: 170px;
-  cursor: pointer;
-`;
-
-const Title = styled.p`
-  margin: 0;
-  font-size: 16px;
-  line-height: 24px;
-  color: rgba(0, 0, 0, 0.85);
-`;
-
-const Paragraph = styled.p`
-  font-size: 14px;
-  line-height: 22px;
-  color: rgba(0, 0, 0, 0.45);
-`;
-
-const ContentWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 0px;
-  gap: 20px;
-  width: 301px;
-  height: 664px;
-`;
