--- plugin/web/extensions/sidebar/popups/TryMapInfo/main.tsx
@@ -1,13 +1,0 @@
-import ReactDOM from "react-dom/client";
-
-import TryMapInfo from ".";
-
-(async () => {
-  const element = document.getElementById("root");
-  if (element) {
-    const root = ReactDOM.createRoot(element);
-    root.render(<TryMapInfo />);
-  }
-})();
-
-export {};
