+++ plugin/web/extensions/sidebar/popups/help/index.tsx
@@ -1,29 +1,26 @@
-import { useState } from "react";
+import { PopupWrapper } from "../sharedComponent";
 
-import { Tab } from "../../core/components/content/Help/hooks";
-import BasicOperation from "../BasicOperation";
-import ClipFunction from "../ClipFunction";
-import ShadowFunction from "../ShadowFunction";
-import TryMapInfo from "../TryMapInfo";
+import BasicOperation from "./BasicOperation";
+import ClipFunction from "./ClipFunction";
+import useGlobalHooks from "./globalHooks";
+import ShadowFunction from "./ShadowFunction";
+import TryMapInfo from "./TryMapInfo";
 
-const Popup: React.FC = () => {
-  const [currentPopup, setCurrentPopup] = useState<Tab>("basic");
+const Help: React.FC = () => {
+  const { currentPopup, handleClosePopup } = useGlobalHooks();
 
-  addEventListener("message", e => {
-    if (e.source !== parent) return null;
-    if (e.data.type) {
-      if (e.data.type === "msgFromHelp" && e.data.message) {
-        setCurrentPopup(e.data.message);
+  return (
+    <PopupWrapper handleClose={handleClosePopup}>
+      {
+        {
+          basic: <BasicOperation />,
+          map: <TryMapInfo />,
+          shadow: <ShadowFunction />,
+          clip: <ClipFunction />,
+        }[currentPopup]
       }
-    }
-  });
-
-  return {
-    basic: <BasicOperation />,
-    map: <TryMapInfo />,
-    shadow: <ShadowFunction />,
-    clip: <ClipFunction />,
-  }[currentPopup];
+    </PopupWrapper>
+  );
 };
 
-export default Popup;
+export default Help;
