+++ plugin/web/sharedComponents/Icon/icons.ts
@@ -37,7 +37,7 @@ import { ReactComponent as Empty } from "./Icons/empty.svg";
 import { ReactComponent as NewPage } from "./Icons/newPage.svg";
 import { ReactComponent as Edit } from "./Icons/edit.svg";
 import { ReactComponent as Bullseye } from "./Icons/bullseye.svg";
-import { ReactComponent as RightArrow } from "./Icons/rightArrow.svg";
+import { ReactComponent as ArrowRight } from "./Icons/arrowRight.svg";
 
 export default {
   logo: Logo,
@@ -76,5 +76,5 @@ export default {
   newPage: NewPage,
   edit: Edit,
   bullseye: Bullseye,
-  rightArrow: RightArrow,
+  arrowRight: ArrowRight,
 };
